Icons: Removed Icons.blend file
This file contains the Grease pencil object of the icons. The rendering method followed is different from usual as it uses Geometry Nodes to change the individual Object index in a collection. A standard design principle is followed which may not be suitable for other developers to follow. Hence, the Icons will only be made by @Tams3d for now. Suggestions are always accepted. This may change in the far future.