Node: Optimize nodes to use the new Index of Nearest
Vertex Slide is ported to the new Index of Nearest node. Does not require Nearest Index. The factor has been remapped to -1 to 1 so that the position of another nearest index can be accessed. Finding another nearest vertex does not work for vertices in the boundary edge as there is no way to offset the nearest index. The factor is unclamped to get another nearest index in a hacky way. In future, a rewrite may require to get a desirable result similar to the Edit mode.